+++ internal/notifier/forwarder_test.go
@@ -18,19 +18,16 @@ package notifier
 
 import (
 	"context"
-	"crypto/x509"
 	"encoding/json"
-	"fmt"
 	"io"
 	"net/http"
 	"net/http/httptest"
 	"testing"
 	"time"
 
-	fuzz "github.com/AdaLogics/go-fuzz-headers"
-	eventv1 "github.com/fluxcd/pkg/apis/event/v1beta1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
+	eventv1 "github.com/fluxcd/pkg/apis/event/v1beta1"
 	"github.com/stretchr/testify/require"
 )
 
@@ -136,34 +133,4 @@ func TestForwarder_Post(t *testing.T) {
 			require.NoError(t, err)
 		})
 	}
-}
-
-func Fuzz_Forwarder(f *testing.F) {
-	f.Add("", []byte{}, []byte{}, []byte{})
-
-	f.Fuzz(func(t *testing.T,
-		urlSuffix string, seed, response, hmacKey []byte) {
-		ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			w.Write(response)
-			io.Copy(io.Discard, r.Body)
-			r.Body.Close()
-		}))
-		defer ts.Close()
-
-		var cert x509.CertPool
-		_ = fuzz.NewConsumer(seed).GenerateStruct(&cert)
-
-		header := make(map[string]string)
-		_ = fuzz.NewConsumer(seed).FuzzMap(&header)
-
-		forwarder, err := NewForwarder(fmt.Sprintf("%s/%s", ts.URL, urlSuffix), "", header, &cert, hmacKey)
-		if err != nil {
-			return
-		}
-
-		event := eventv1.Event{}
-		_ = fuzz.NewConsumer(seed).GenerateStruct(&event)
-
-		_ = forwarder.Post(context.TODO(), event)
-	})
 }
