+++ internal/notifier/github_test.go
@@ -17,16 +17,8 @@ limitations under the License.
 package notifier
 
 import (
-	"context"
-	"crypto/x509"
-	"fmt"
-	"io"
-	"net/http"
-	"net/http/httptest"
 	"testing"
 
-	fuzz "github.com/AdaLogics/go-fuzz-headers"
-	eventv1 "github.com/fluxcd/pkg/apis/event/v1beta1"
 	"github.com/google/go-github/v41/github"
 	"github.com/stretchr/testify/assert"
 )
@@ -75,47 +67,6 @@ func TestDuplicateGithubStatus(t *testing.T) {
 	for _, test := range tests {
 		assert.Equal(test.dup, duplicateGithubStatus(test.ss, test.s))
 	}
-}
-
-func Fuzz_GitHub(f *testing.F) {
-	f.Add("token", "org/repo", "revision/abce1", "error", "", []byte{}, []byte(`[{"context":"/","state":"failure","description":""}]`))
-	f.Add("token", "org/repo", "revision/abce1", "info", "", []byte{}, []byte(`[{"context":"/","state":"success","description":""}]`))
-	f.Add("token", "org/repo", "revision/abce1", "info", "", []byte{}, []byte(`[{"context":"/","state":"failure","description":""}]`))
-	f.Add("token", "org/repo", "revision/abce1", "info", "", []byte{}, []byte(`[{"context":"/"}]`))
-	f.Add("token", "org/repo", "revision/abce1", "info", "", []byte{}, []byte(`[{}]`))
-	f.Add("token", "org/repo", "revision/abce1", "info", "Progressing", []byte{}, []byte{})
-
-	f.Fuzz(func(t *testing.T,
-		token, urlSuffix, revision, severity, reason string, seed, response []byte) {
-		ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			w.Write(response)
-			io.Copy(io.Discard, r.Body)
-			r.Body.Close()
-		}))
-		defer ts.Close()
-
-		var cert x509.CertPool
-		_ = fuzz.NewConsumer(seed).GenerateStruct(&cert)
-
-		github, err := NewGitHub(fmt.Sprintf("%s/%s", ts.URL, urlSuffix), token, &cert)
-		if err != nil {
-			return
-		}
-
-		event := eventv1.Event{}
-		_ = fuzz.NewConsumer(seed).GenerateStruct(&event)
-
-		if event.Metadata == nil && (revision != "") {
-			event.Metadata = map[string]string{
-				"revision": revision,
-			}
-		}
-
-		event.Severity = severity
-		event.Reason = reason
-
-		_ = github.Post(context.TODO(), event)
-	})
 }
 
 func ghStatus(state string, context string, description string) *github.RepoStatus {
