+++ internal/notifier/rocket_test.go
@@ -18,16 +18,12 @@ package notifier
 
 import (
 	"context"
-	"crypto/x509"
 	"encoding/json"
-	"fmt"
 	"io"
 	"net/http"
 	"net/http/httptest"
 	"testing"
 
-	fuzz "github.com/AdaLogics/go-fuzz-headers"
-	eventv1 "github.com/fluxcd/pkg/apis/event/v1beta1"
 	"github.com/stretchr/testify/require"
 )
 
@@ -49,40 +45,4 @@ func TestRocket_Post(t *testing.T) {
 
 	err = rocket.Post(context.TODO(), testEvent())
 	require.NoError(t, err)
-}
-
-func Fuzz_Rocket(f *testing.F) {
-	f.Add("user", "channel", "", "error", "", "", []byte{}, []byte{})
-	f.Add("user", "channel", "", "error", "update", "", []byte{}, []byte{})
-
-	f.Fuzz(func(t *testing.T,
-		username, channel, urlSuffix, severity, commitStatus, message string, seed, response []byte) {
-		ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			w.Write(response)
-			io.Copy(io.Discard, r.Body)
-			r.Body.Close()
-		}))
-		defer ts.Close()
-
-		var cert x509.CertPool
-		_ = fuzz.NewConsumer(seed).GenerateStruct(&cert)
-
-		rocket, err := NewRocket(fmt.Sprintf("%s/%s", ts.URL, urlSuffix), "", &cert, username, channel)
-		if err != nil {
-			return
-		}
-
-		event := eventv1.Event{}
-		_ = fuzz.NewConsumer(seed).GenerateStruct(&event)
-
-		if event.Metadata == nil {
-			event.Metadata = map[string]string{}
-		}
-
-		event.Metadata["commit_status"] = commitStatus
-		event.Severity = severity
-		event.Message = message
-
-		_ = rocket.Post(context.TODO(), event)
-	})
 }
