+++ internal/notifier/util_test.go
@@ -19,7 +19,6 @@ package notifier
 import (
 	"testing"
 
-	fuzz "github.com/AdaLogics/go-fuzz-headers"
 	eventv1 "github.com/fluxcd/pkg/apis/event/v1beta1"
 	"github.com/stretchr/testify/require"
 	v1 "k8s.io/api/core/v1"
@@ -123,25 +122,4 @@ func TestUtil_BasicAuth(t *testing.T) {
 	password := "password"
 	s := basicAuth(username, password)
 	require.Equal(t, "dXNlcjpwYXNzd29yZA==", s)
-}
-
-func Fuzz_Util_ParseGitAddress(f *testing.F) {
-	f.Add("ssh://git@abc.com")
-
-	f.Fuzz(func(t *testing.T, gitAddress string) {
-		_, _, _ = parseGitAddress(gitAddress)
-	})
-}
-
-func Fuzz_Util_FormatNameAndDescription(f *testing.F) {
-	f.Add("aA1-", []byte{})
-
-	f.Fuzz(func(t *testing.T, reason string, seed []byte) {
-		event := eventv1.Event{}
-		_ = fuzz.NewConsumer(seed).GenerateStruct(&event)
-
-		event.Reason = reason
-
-		_, _ = formatNameAndDescription(event)
-	})
 }
